introduce warnings in spec
