+++ packages/provider/src/image-model/v1/image-model-v1.ts
@@ -1,4 +1,5 @@
-import { JSONValue } from '../../json-value/json-value';
+import { ImageModelV1CallOptions } from './image-model-v1-call-options';
+import { ImageModelV1Warning } from './image-model-v1-call-warning';
 
 /**
 Image generation model specification version 1.
@@ -32,64 +33,7 @@ If undefined, we will max generate one image per call.
   /**
 Generates an array of images.
    */
-  doGenerate(options: {
-    /**
-Prompt for the image generation.
-     */
-    prompt: string;
-
-    /**
-Number of images to generate.
-     */
-    n: number;
-
-    /**
-Size of the images to generate.
-Must have the format `{width}x{height}`.
-`undefined` will use the provider's default size.
-     */
-    size: `${number}x${number}` | undefined;
-
-    /**
-Aspect ratio of the images to generate.
-Must have the format `{width}:{height}`.
-`undefined` will use the provider's default aspect ratio.
-     */
-    aspectRatio: `${number}:${number}` | undefined;
-
-    /**
-Seed for the image generation.
-`undefined` will use the provider's default seed.
-     */
-    seed: number | undefined;
-
-    /**
-Additional provider-specific options that are passed through to the provider
-as body parameters.
-
-The outer record is keyed by the provider name, and the inner
-record is keyed by the provider-specific metadata key.
-```ts
-{
-  "openai": {
-    "style": "vivid"
-  }
-}
-```
-     */
-    providerOptions: Record<string, Record<string, JSONValue>>;
-
-    /**
-Abort signal for cancelling the operation.
-     */
-    abortSignal?: AbortSignal;
-
-    /**
-  Additional HTTP headers to be sent with the request.
-  Only applicable for HTTP-based providers.
-     */
-    headers?: Record<string, string | undefined>;
-  }): PromiseLike<{
+  doGenerate(options: ImageModelV1CallOptions): PromiseLike<{
     /**
 Generated images as base64 encoded strings or binary data.
 The images should be returned without any unnecessary conversion.
@@ -98,5 +42,10 @@ as base64 encoded strings. If the API returns binary data, the images should
 be returned as binary data.
      */
     images: Array<string> | Array<Uint8Array>;
+
+    /**
+Warnings for the call, e.g. unsupported settings.
+     */
+    warnings?: Array<ImageModelV1Warning>;
   }>;
 };
