+++ packages/provider/src/image-model/v1/index.ts
@@ -1,1 +1,2 @@
-export * from './image-model-v1';
+export type { ImageModelV1 } from './image-model-v1';
+export type { ImageModelV1CallOptions } from './image-model-v1-call-options';
